app/pyproject.toml
[tool.poetry]
name = "grimeguardian"
version = "0.1.0"
description = ""
authors = ["Sunil Kumar <sdkumar@hmc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
groundlight = "^0.13.1"
python-dotenv = "^1.0.0"
yappi = "^1.6.0"
certifi = ">=2023.7.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
